% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IFconversion.r
\docType{data}
\name{data_IF}
\alias{data_IF}
\title{The sample intuitionistic fuzzy dataset}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 5 rows and 9 columns.
}
\usage{
data_IF
}
\description{
The sample intuitionistic fuzzy dataset
}
\examples{
set.seed(61222)
data(data_IF)
m<-IFSM(data_IF)
print(m)
}
\keyword{data}
